+++ SDLGameEngine/TextureManager.h
@@ -26,7 +26,7 @@ public:
     
     bool load(string fileName, string id, SDL_Renderer* pRenderer);
     
-    void draw(string id, int x, int y,
+    void draw(string id, float x, float y,
               int width, int height,
               SDL_Renderer* pRenderer, SDL_RendererFlip flip = SDL_FLIP_NONE);
     
@@ -35,7 +35,7 @@ public:
               SDL_Renderer* pRenderer, SDL_RendererFlip flip);
 
     
-    void drawFrame(string id, int x, int y,
+    void drawFrame(string id, float x, float y,
                    int width, int height,int currentRow, int currentFrame,
                    float zoom ,SDL_Renderer* pRenderer,
                    SDL_RendererFlip flip = SDL_FLIP_NONE);
@@ -50,8 +50,8 @@ public:
                    int currentRow, int currentFrame,
                    SDL_Renderer* pRenderer, SDL_RendererFlip flip);
     
-    void drawTile(string id, int margin, int spacing, int x,
-                  int y, int width, int height,int currentRow,
+    void drawTile(string id, int margin, int spacing, float x,
+                  float y, int width, int height,int currentRow,
                   int currentFrame, float zoom, SDL_Renderer* pRenderer);
